+++ packages/react-native-sdk/__tests__/components/CallControls.test.tsx
@@ -4,7 +4,12 @@ import mockParticipant from '../mocks/participant';
 import { ButtonTestIds, ComponentTestIds } from '../../src/constants/TestIds';
 import { mockCall } from '../mocks/call';
 import { fireEvent, render, screen, waitFor } from '../utils/RNTLTools';
-import { CallControls } from '../../src/components';
+import {
+  CallControls,
+  ChatButton,
+  HangUpCallButton,
+  ReactionButton,
+} from '../../src/components';
 import { OwnCapability } from '@stream-io/video-client';
 
 console.warn = jest.fn();
@@ -14,7 +19,7 @@ enum P_IDS {
   LOCAL_1 = 'local-1',
 }
 
-describe('CallControls', () => {
+describe('ChatButton', () => {
   it('should render an unread badge indicator when the value is defined in the chatButton prop', async () => {
     const call = mockCall(mockClientWithUser(), [
       mockParticipant({
@@ -24,14 +29,9 @@ describe('CallControls', () => {
       }),
     ]);
 
-    render(
-      <CallControls
-        chatButton={{ onPressHandler: jest.fn(), unreadBadgeCountIndicator: 1 }}
-      />,
-      {
-        call,
-      },
-    );
+    render(<ChatButton onPressHandler={jest.fn()} unreadBadgeCount={1} />, {
+      call,
+    });
 
     const indicator = await screen.findByText('1');
 
@@ -47,14 +47,9 @@ describe('CallControls', () => {
       }),
     ]);
 
-    render(
-      <CallControls
-        chatButton={{ onPressHandler: jest.fn(), unreadBadgeCountIndicator: 0 }}
-      />,
-      {
-        call,
-      },
-    );
+    render(<ChatButton onPressHandler={jest.fn()} unreadBadgeCount={0} />, {
+      call,
+    });
 
     await waitFor(() =>
       expect(() =>
@@ -64,7 +59,9 @@ describe('CallControls', () => {
       ),
     );
   });
+});
 
+describe('ReactionButton', () => {
   it('render reaction button in call controls component', async () => {
     const call = mockCall(
       mockClientWithUser(),
@@ -78,7 +75,7 @@ describe('CallControls', () => {
       [OwnCapability.CREATE_REACTION],
     );
 
-    render(<CallControls />, {
+    render(<ReactionButton />, {
       call,
     });
 
@@ -88,7 +85,9 @@ describe('CallControls', () => {
 
     expect(screen.getByTestId(ComponentTestIds.REACTIONS_PICKER)).toBeVisible();
   });
+});
 
+describe('HangupCallButton', () => {
   it('execute onPressHandler when its passed to hangup call button when its pressed in call controls component', async () => {
     const call = mockCall(mockClientWithUser(), [
       mockParticipant({
@@ -98,9 +97,9 @@ describe('CallControls', () => {
       }),
     ]);
 
-    const hangUpCallButton = { onPressHandler: jest.fn() };
+    const onHangupCallHandler = jest.fn();
 
-    render(<CallControls hangUpCallButton={hangUpCallButton} />, {
+    render(<HangUpCallButton onPressHandler={onHangupCallHandler} />, {
       call,
     });
 
@@ -108,7 +107,7 @@ describe('CallControls', () => {
 
     fireEvent.press(button);
 
-    expect(hangUpCallButton.onPressHandler).toHaveBeenCalled();
+    expect(onHangupCallHandler).toHaveBeenCalled();
   });
 
   it('execute call.leave when hangup button is pressed with no custom handler in call controls component', async () => {