+++ packages/react-native-sdk/src/components/call/index.ts
@@ -1,6 +1,0 @@
-export * from './CallContent';
-export * from './CallControls';
-export * from './CallParticipantsList';
-export * from './IncomingCall';
-export * from './OutgoingCall';
-export * from './Lobby';