+++ packages/react-native-sdk/src/components/participants/index.ts
@@ -1,4 +1,0 @@
-export * from './LocalParticipantView';
-export * from './ParticipantsInfoBadge';
-export * from './Participant';
-export * from './ParticipantsInfoList';