+++ sample-apps/react/react-video-demo/src/components/JoinContainer/JoinContainer.tsx
@@ -4,24 +4,23 @@ import classnames from 'classnames';
 import Button from '../Button';
 
 import styles from './JoinContainer.module.css';
+import { logoURI } from '../../utils/constants';
 
 export type Props = {
   className?: string;
   joinCall(): void;
-  logo: string;
   isJoiningCall?: boolean;
 };
 
 export const JoinContainer: FC<Props> = ({
   className,
   joinCall,
-  logo,
   isJoiningCall,
 }) => {
   const rootClassName = classnames(styles.root, className);
   return (
     <div className={rootClassName}>
-      <img src={logo} className={styles.logo} />
+      <img src={logoURI} className={styles.logo} />
       <p className={styles.description}>
         You are about to start a private test call via Stream. Once you start
         the call, you can invite other participants